requirements.txt
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.1
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==2.6
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.18.0
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
